move utils package
