+++ songsuro/autoencoder/decoder/generator.py
@@ -3,7 +3,7 @@ import torch.nn.functional as F
 import torch.nn as nn
 from torch.nn import Conv1d, ConvTranspose1d
 from torch.nn.utils import weight_norm, remove_weight_norm
-from songsuro.autoencoder.decoder.utils import get_padding, init_weights
+from songsuro.utils.util import get_padding, init_weights
 
 LRELU_SLOPE = 0.1
