+++ songsuro/utils/util.py
@@ -53,3 +53,13 @@ def temporal_avg_pool(x, mask=None):
 	out = torch.div(x, len_)
 	out = torch.where(zero_mask, torch.zeros_like(out), out)
 	return out
+
+
+def init_weights(m, mean=0.0, std=0.01):
+	classname = m.__class__.__name__
+	if classname.find("Conv") != -1:
+		m.weight.data.normal_(mean, std)
+
+
+def get_padding(kernel_size, dilation=1):
+	return int((kernel_size * dilation - dilation) / 2)
